refactor: Refactor main.js to improve code readability and reusable component
