+++ index.html
@@ -8,27 +8,7 @@
     <link rel="stylesheet" href="/styles.css" />
   </head>
   <body>
-    <div class="weather-info-card">
-      <div class="weather-location"></div>
-      <div class="weather-time"></div>
-      <div class="weather-details">
-        <div class="weather-summary">
-          <img class="weather-summary-icon"></img>  
-          <div class="weather-summary-description"></div>
-        </div>
-        <div class="weather-temperature-info">
-          <div class="temperature-current">
-            <div class="temperature-current-label">현재</div>
-            <div class="temperature-current-value"></div>
-          </div>
-          <div class="temperature-min-max">
-            <div class="temperature-min-max-label">최저/최고</div>
-            <div class="temperature-min-max-value"></div>
-            
-          </div>
-        </div>
-      </div>
-    </div>
     <script type="module" src="/main.js"></script>
+    <script type="module" src="/detail.js"></script>
   </body>
 </html>
