+++ utils/date.js
@@ -1,0 +1,3 @@
+export const formatDate = (timestamp, options) => {
+  return new Date(timestamp * 1000).toLocaleString('ko-KR', options);
+};
